Automatically set aws region when using iam as an auth method (#6)
Currently the AWS client does not know which endpoint to call when using IAM. You have to set the AWS default region in the Kubedb operator docker container, either via the environment variable, or through config. This is a really bizarre thing to do, and causes this error https://github.com/kubedb/project/issues/210. If you don't understand the architecture fully or are new to Kubedb, its hard to understand why and how this is happening. 

Ideally, the AWS client would figure out the clients region automagically from the API. There is an open tracked issue for this here https://github.com/aws/aws-sdk-go/issues/1103

I didn't manage to test this and i didn't try to fix the hardcoded values below.
Any hints on testing this would be great.